backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-decouple
pydantic
pydantic-settings
onnxruntime
opencv-python
pillow
numpy
scipy
scikit-learn
pandas
python-logging-json-formatter
starlette-admin
httpx
pytest
pytest-asyncio
pytest-cov
black
isort
flake8
mypy
aiofiles
python-magic
python-dotenv
gunicorn
